+++ apps/marketing/src/app/layout.tsx
@@ -1,10 +1,12 @@
 import '@/styles/globals.css'
 import { cms } from '@/cms'
 import { AppLayout } from '@/components/app-layout'
+import { TooltipProvider } from '@/components/ui/tooltip'
+import { WorkbenchTitle } from '@/components/workbench-title'
+import { PostHogProvider } from '@/posthog'
+import { ClerkProvider } from '@clerk/nextjs'
 import { DeepDishProvider } from '@deepdish/cms'
 import { GeistMono } from 'geist/font/mono'
-
-import { ClerkProvider } from '@clerk/nextjs'
 
 export const metadata = {
   title: 'Manage content directly on your pages',
@@ -39,9 +41,13 @@ export default function RootLayout({
     <ClerkProvider>
       <html lang="en" className="scroll-smooth">
         <body className={`${GeistMono.className} antialiased`}>
-          <DeepDishProvider title="DeepDish Example Workbench">
-            <AppLayout>{children}</AppLayout>
-          </DeepDishProvider>
+          <TooltipProvider>
+            <DeepDishProvider title={<WorkbenchTitle />} authDisabled={true}>
+              <PostHogProvider>
+                <AppLayout>{children}</AppLayout>
+              </PostHogProvider>
+            </DeepDishProvider>
+          </TooltipProvider>
         </body>
       </html>
     </ClerkProvider>