+++ packages/cms/src/provider/container.tsx
@@ -1,12 +1,20 @@
 import { getSettings } from '@deepdish/ui/config'
 import { Provider } from './provider'
 
-export function Settings(props: { children: React.ReactNode; title?: string }) {
+export function ProviderContainer(props: {
+  children: React.ReactNode
+  title?: React.ReactNode
+  authDisabled?: boolean
+}) {
   const settings = getSettings()
   const draft = settings.failure ? false : settings.data.draft
 
   return (
-    <Provider draft={draft} title={props.title}>
+    <Provider
+      draft={draft}
+      title={props.title}
+      authDisabled={props.authDisabled}
+    >
       {props.children}
     </Provider>
   )