+++ packages/workbench/lib/index.tsx
@@ -1,17 +1,26 @@
 'use client'
 
 import '../src/index.css'
-import { Workbench as InternalWorkbench } from '@/components/workbench'
+import {
+  Workbench as InternalWorkbench,
+  type WorkbenchProps as InternalWorkbenchProps,
+} from '@/components/workbench'
 import { useRef } from 'react'
 import { Tailwind } from 'react-shadow-scope'
 
-export function Workbench({ title }: { title?: string }) {
+type WorkbenchProps = Omit<InternalWorkbenchProps, 'ref'>
+
+export function Workbench(props: WorkbenchProps) {
   const ref = useRef<HTMLDivElement>(null)
 
   return (
     <Tailwind href="/__deepdish/css">
       <div ref={ref} className="deepdish-shadow">
-        <InternalWorkbench ref={ref} title={title} />
+        <InternalWorkbench
+          ref={ref}
+          title={props.title}
+          authDisabled={props.authDisabled}
+        />
       </div>
     </Tailwind>
   )