+++ packages/cms/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @deepdish/cms
 
+## 0.3.0
+
+### Minor Changes
+
+- 9e94ae0: Added `authDisabled` flag and refactored the `title` prop.
+
+### Patch Changes
+
+- 47e41fe: Collocated the provider and container.
+- Updated dependencies [9e94ae0]
+- Updated dependencies [9e94ae0]
+  - @deepdish/nextjs@0.4.0
+  - @deepdish/workbench@0.8.0
+
+## 0.2.0
+
+### Minor Changes
+
+- 1208209: Implemented core components of cms package.
+
+### Patch Changes
+
+- 8787ed4: Conditionally rendered workbench based on draft setting.
+- Updated dependencies [edd3954]
+- Updated dependencies [142e359]
+  - @deepdish/ui@0.11.1
+  - @deepdish/core@0.1.1
+  - @deepdish/workbench@0.7.1
+  - @deepdish/nextjs@0.3.1
+
 ## 0.1.1
 
 ### Patch Changes