+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @deepdish/nextjs
 
+## 0.4.0
+
+### Minor Changes
+
+- 9e94ae0: Added the `health` middleware option.
+
+### Patch Changes
+
+- Updated dependencies [9e94ae0]
+  - @deepdish/workbench@0.8.0
+
+## 0.3.1
+
+### Patch Changes
+
+- @deepdish/workbench@0.7.1
+
 ## 0.3.0
 
 ### Minor Changes